feat: 重構目的地設定檔管理邏輯
- 將原有的 DestinationProfileService 拆分為 ProfileManager 和 ProfileResolver，提供更清晰的職責分離。
- ProfileManager 負責管理目的地設定檔的 CRUD 操作，並處理與使用者帳戶相關的權限檢查。
- ProfileResolver 專注於解決與儲存相關的設定檔，確保在儲存過程中正確處理預設設定檔的遷移。
- 新增 ProfileStore.js，集中管理目的地設定檔的存取邏輯與錯誤處理。
- 更新單元測試以涵蓋新的 ProfileManager 和 ProfileResolver 的行為，確保功能正確性。

這些變更旨在提高代碼的可維護性和可讀性，並為未來的擴展提供更好的基礎。
