+++ tools/package-extension.sh
@@ -117,6 +117,7 @@ rsync -a \
     --exclude='config/ui-selectors.js' \
     --exclude='config/README.md' \
     --exclude='content' \
+    --exclude='destinations/ProfileResolver.js' \
     --exclude='config/env/build.example.js' \
     --exclude='config/env.js' \
     --exclude='config/extension/index.js' \
@@ -146,10 +147,6 @@ rsync -a \
 # Sidepanel 直接依賴 HighlightLookupResolver，需顯式補回 package。
 mkdir -p "$RM_DIR/scripts/highlighter/core"
 cp -a scripts/highlighter/core/HighlightLookupResolver.js "$RM_DIR/scripts/highlighter/core/"
-
-# Popup / Options 直接依賴 destination profile service，需顯式補回 package。
-mkdir -p "$RM_DIR/scripts/background/services"
-cp -a scripts/background/services/DestinationProfileService.js "$RM_DIR/scripts/background/services/"
 
 # 清理 macOS metadata，避免從 cp -a 帶入 release package。
 find "$RM_DIR" -name '.DS_Store' -delete
